wip: refactor of the lua stuff.
